+++ pakages/cache.py
@@ -106,8 +106,13 @@ class BuildCache:
 
             package_name = os.path.basename(archive)
             full_name = "%s/%s:%s" % (uri, package_name, tag)
-            oras.push(full_name, archive, content_type=content_type)
 
+            # Push a full hash and a generic one
+            oras.push(full_name, archive, content_type=content_type)
+            generic_name = utils.generalize_spack_archive(package_name)
+            uri = "%s/%s:%s" % (uri, generic_name, tag)
+            full_name = "%s/%s:%s" % (uri, generic_name, tag)
+            oras.push(full_name, archive, content_type=content_type)
             # TODO how to add sbom? separately?
             # TODO we should support custom tags