+++ pakages/cli/push.py
@@ -3,7 +3,6 @@ __copyright__ = "Copyright 2021-2022, Vanessa Sochat and Alec Scott"
 __license__ = "Apache-2.0"
 
 from pakages.client import PakClient
-from pakages.logger import logger
 
 
 def main(args, parser, extra, subparser):