+++ pakages/utils/__init__.py
@@ -6,7 +6,12 @@ from .terminal import (
     which,
     confirm_action,
 )
-from .spack import add_spack_to_path, add_pakages_spack_repo, install_spack
+from .spack import (
+    add_spack_to_path,
+    add_pakages_spack_repo,
+    install_spack,
+    generalize_spack_archive,
+)
 from .fileio import (
     copyfile,
     get_file_hash,